moved utils methods from Parser to Utils
